# fitnessWebApp-UI/src/app/services/models/progresso-response.ts
/* tslint:disable */
/* eslint-disable */
export interface ProgressoResponse {
  altezzaCm: number;
  dataMisurazione: string;
  id: number;
  note?: string;
  percentualeMassaGrassa: number;
  percentualeMassaMagra: number;
  pesoKg: number;
}