# fitnessWebApp-UI/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Andrea Barutta",
    "description": "OpenApi documentation for Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Andrea Barutta",
      "url": "https://barutta02.github.io/Barutta_portfolio/",
      "email": "andreabarutta123@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Local ENV"
    },
    {
      "url": "https://some-url.com/api/v1",
      "description": "PROD ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/progress/{progresso-id}": {
      "get": {
        "tags": [
          "Progress"
        ],
        "operationId": "getProgresso",
        "parameters": [
          {
            "name": "progresso-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressoResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Progress"
        ],
        "operationId": "updateProgresso",
        "parameters": [
          {
            "name": "progresso-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Progress"
        ],
        "operationId": "deleteProgresso",
        "parameters": [
          {
            "name": "progresso-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/progress": {
      "post": {
        "tags": [
          "Progress"
        ],
        "operationId": "addProgresso",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressoResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/periodTraining": {
      "post": {
        "tags": [
          "PeriodTraining"
        ],
        "operationId": "savePeriodoAllenamento",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeriodoAllenamentoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeriodoAllenamentoResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/progress/last/{N}": {
      "get": {
        "tags": [
          "Progress"
        ],
        "operationId": "getLastNProgressi",
        "parameters": [
          {
            "name": "N",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgressoResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/periods/{periodo-id}": {
      "get": {
        "tags": [
          "Periods"
        ],
        "operationId": "findPeriodoById",
        "parameters": [
          {
            "name": "periodo-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeriodoResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/periodTraining/{periodo-nome}": {
      "get": {
        "tags": [
          "PeriodTraining"
        ],
        "operationId": "findByPeriodoNome",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "periodo-nome",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponsePeriodoAllenamentoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/periodTraining/no_pagination/{periodo-nome}": {
      "get": {
        "tags": [
          "PeriodTraining"
        ],
        "operationId": "findByPeriodoNomeNoPagination",
        "parameters": [
          {
            "name": "periodo-nome",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PeriodoAllenamentoResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/periods/{periodo-nome}": {
      "delete": {
        "tags": [
          "Periods"
        ],
        "operationId": "deletePeriodo",
        "parameters": [
          {
            "name": "periodo-nome",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/periodTraining/{periodo-allenamento-id}": {
      "delete": {
        "tags": [
          "PeriodTraining"
        ],
        "operationId": "deletePeriodoAllenamento",
        "parameters": [
          {
            "name": "periodo-allenamento-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProgressoRequest": {
        "required": [
          "altezzaCm",
          "dataMisurazione",
          "percentualeMassaGrassa",
          "percentualeMassaMagra",
          "pesoKg"
        ],
        "type": "object",
        "properties": {
          "dataMisurazione": {
            "type": "string",
            "format": "date"
          },
          "pesoKg": {
            "type": "number",
            "format": "float"
          },
          "altezzaCm": {
            "type": "number",
            "format": "float"
          },
          "percentualeMassaGrassa": {
            "type": "number",
            "format": "float"
          },
          "percentualeMassaMagra": {
            "type": "number",
            "format": "float"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "PeriodoResponse": {
        "required": [
          "attivo",
          "data_inizio",
          "durata_in_giorni",
          "id",
          "name",
          "obiettivo"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "obiettivo": {
            "type": "string",
            "enum": [
              "DIMAGRIMENTO",
              "DEFINIZIONE",
              "MASSA",
              "FORZA",
              "RESISTENZA",
              "NON_DEFINITO"
            ]
          },
          "durata_in_giorni": {
            "type": "integer",
            "format": "int32"
          },
          "data_inizio": {
            "type": "string",
            "format": "date"
          },
          "data_fine": {
            "type": "string",
            "format": "date"
          },
          "attivo": {
            "type": "boolean"
          },
          "creator_username": {
            "type": "string"
          }
        }
      },
...
      "ProgressoResponse": {
        "required": [
          "altezzaCm",
          "dataMisurazione",
          "id",
          "percentualeMassaGrassa",
          "percentualeMassaMagra",
          "pesoKg"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "dataMisurazione": {
            "type": "string",
            "format": "date"
          },
          "pesoKg": {
            "type": "number",
            "format": "float"
          },
          "altezzaCm": {
            "type": "number",
            "format": "float"
          },
          "percentualeMassaGrassa": {
            "type": "number",
            "format": "float"
          },
          "percentualeMassaMagra": {
            "type": "number",
            "format": "float"
          },
          "note": {
            "type": "string"
          }
        }
      },
...
      "PeriodoAllenamentoRequest": {
        "required": [
          "giorno_del_periodo",
          "id_allenamento",
          "id_periodo",
          "periodo_giornata"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "id_allenamento": {
            "type": "integer",
            "format": "int64"
          },
          "id_periodo": {
            "type": "integer",
            "format": "int64"
          },
          "giorno_del_periodo": {
            "type": "integer",
            "format": "int32"
          },
          "periodo_giornata": {
            "type": "string",
            "enum": [
              "MATTINA",
              "POMERIGGIO",
              "SERA"
            ]
          }
        }
      },
      "PeriodoAllenamentoResponse": {
        "required": [
          "giorno_del_periodo",
          "id",
          "id_allenamento",
          "id_periodo",
          "periodo_giornata"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "id_allenamento": {
            "type": "integer",
            "format": "int64"
          },
          "id_periodo": {
            "type": "integer",
            "format": "int64"
          },
          "giorno_del_periodo": {
            "type": "integer",
            "format": "int32"
          },
          "periodo_giornata": {
            "type": "string",
            "enum": [
              "MATTINA",
              "POMERIGGIO",
              "SERA"
            ]
          }
        }
      },
...
      "PageResponsePeriodoAllenamentoResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PeriodoAllenamentoResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}